refactor(Trigger): add abstract class base to provide same method to extend child classes
